refactor(announcements): React-free engine types, and the re-points they force
v3 Phase 3, Task 3.1. `src/lib/announcements-engine/types.ts` holds the type
surface the engine barrel reaches: `AnnouncementState`, `EngineAnnouncementConfig`,
the queue types and `DEFAULT_QUEUE_CONFIG`, the action union and `IsScheduleActive`.

The dependency direction is the point. `types/announcement.ts` now imports FROM
the engine and widens `title`/`description` to `ReactNode`; the engine never
learns those fields exist. `AnnouncementConfig extends EngineAnnouncementConfig`
and re-declares `variant` as required — legal narrowing, and it lets an
engine-only consumer register `{ id }` without picking a presentation variant.

The plan's Decision 3 listed eight re-points. The test plan's §0 C2 measured
eleven, and it is right: EVERY engine-reachable `core/*` leaf type-imported
`../types/announcement`, whose first line is `import type { ReactNode } from
'react'`, and `types/queue.ts` reached it too. A type import leaves no trace in
the emitted JS, so `dist:size` and a JS closure scan both stay green while
`react` sits in the declaration chain (recipe gap 18) — only the source walk
sees it. `core/{frequency,priority-queue,scheduler}.ts` and `types/queue.ts` now
take their types from the engine file.

Also re-pointed, all 0 B: `core/audience.ts` and `adapters/sonner.ts` to
`@tour-kit/core/engine`, and `core/resolve-schedule.ts` plus the config's
`schedule` field to `@tour-kit/scheduling/engine` — the main entry's `.d.ts`
names `react` four times (Decision 7b item 2). The runtime `require(
'@tour-kit/scheduling')` at resolve-schedule.ts:33 stays: it is the optional
peer load, measured-present rather than forbidden.

`DismissalReason` in the first draft of the engine types was invented; the real
union is the seven-member one this moves verbatim.

setup.tsx grows one line, landed here on its own so a flip would be a finding:
the localStorage mock is module scope and was never cleared, so a blob written
by one case was read by the next. 41/377 green against it alone — no existing
test depended on the leak.

Oracle: 41 files / 377 passed. typecheck clean.
